test(dedupe): consolidate logging patterns tests (8 → 3 files)
Consolidate 8 test_logging_patterns_* files into 3 organized files:
- test_logging_patterns.py (core: StructuredLogger, log_operation, log_execution)
- test_logging_patterns_domain.py (trades, positions, market data)
- test_logging_patterns_health.py (system health, config, errors)

All files ≤240 lines per dedupe guidelines.
Cluster: 2a83957b1884
