+++ tests/unit/gpt_trader/utilities/test_logging_patterns.py
@@ -1,9 +1,80 @@
+"""Tests for core logging patterns: StructuredLogger, log_operation, log_execution."""
+
 from __future__ import annotations
 
+import logging
 from contextlib import contextmanager
 
 from gpt_trader.utilities import logging_patterns as lp
-from tests.unit.gpt_trader.utilities.logging_patterns_test_helpers import captured_logger
+from tests.unit.gpt_trader.utilities.logging_patterns_test_helpers import (
+    captured_logger,
+    has_attr,
+    messages,
+)
+
+# ============================================================================
+# StructuredLogger Tests
+# ============================================================================
+
+
+def test_structured_logger_methods_apply_component():
+    with captured_logger("structured") as (_, handler):
+        slog = lp.StructuredLogger("structured", component="orders")
+        slog.info("info message", operation="test")
+        slog.warning("warn message", operation="test")
+        slog.error("error message", operation="test")
+        slog.debug("debug message", operation="test")
+        slog.critical("critical message", operation="test")
+        slog.log(logging.INFO, "generic message", operation="test")
+
+    assert len(handler.records) == 6
+    assert all(has_attr(rec, "component", "orders") for rec in handler.records)
+    assert handler.records[0].getMessage().startswith("info message")
+
+
+def test_get_logger_returns_structured_logger():
+    slog = lp.get_logger("custom", component="risk")
+    assert isinstance(slog, lp.StructuredLogger)
+    with captured_logger("custom") as (_, handler):
+        slog.info("hello", operation="greet")
+    record = handler.records[0]
+    assert has_attr(record, "operation", "greet")
+
+
+# ============================================================================
+# log_operation Tests
+# ============================================================================
+
+
+def test_log_operation_with_plain_logger():
+    with captured_logger("operation") as (logger, handler):
+        with lp.log_operation("fetch", logger=logger, symbol="BTC"):
+            pass
+
+    start_msg, end_msg = messages(handler)
+    start_rec, end_rec = handler.records
+
+    assert "Started fetch" in start_msg
+    assert has_attr(start_rec, "operation", "fetch")
+    assert has_attr(start_rec, "symbol", "BTC")
+    assert "Completed fetch" in end_msg
+    assert has_attr(end_rec, "duration_ms")
+    assert handler.records[0].name == "operation"
+
+
+def test_log_operation_with_default_logger():
+    with captured_logger("operation") as (_, handler):
+        with lp.log_operation("sync"):
+            pass
+
+    start, end = messages(handler)
+    assert "Started sync" in start
+    assert "Completed sync" in end
+
+
+# ============================================================================
+# log_execution Decorator Tests
+# ============================================================================
 
 
 def test_log_execution_includes_args_and_result():
