+++ tests/unit/gpt_trader/utilities/test_logging_patterns_domain.py
@@ -1,6 +1,7 @@
+"""Tests for domain-specific logging: trades, positions, market data."""
+
 from __future__ import annotations
 
-import logging
 from decimal import Decimal
 from typing import Any
 
@@ -9,6 +10,10 @@ from tests.unit.gpt_trader.utilities.logging_patterns_test_helpers import (
     captured_logger,
     has_attr,
 )
+
+# ============================================================================
+# Trade Event Tests
+# ============================================================================
 
 
 def test_log_trade_event_includes_all_context():
@@ -52,6 +57,11 @@ def test_log_trade_event_with_structured_logger():
     assert has_attr(record, "component", "desk")
 
 
+# ============================================================================
+# Position Update Tests
+# ============================================================================
+
+
 def test_log_position_update_adds_metrics():
     with captured_logger("position") as (logger, handler):
         lp.log_position_update(
@@ -88,17 +98,6 @@ def test_log_position_update_with_plain_logger():
     assert has_attr(record, "operation", "position_update")
 
 
-def test_log_position_update_plain_logger_branch(caplog):
-    caplog.set_level(logging.INFO, logger="position-branch")
-    logger = logging.getLogger("position-branch")
-    lp.log_position_update("SOL", position_size=Decimal("3"), logger=logger)
-
-    assert any(
-        rec.name == "position-branch" and has_attr(rec, "operation", "position_update")
-        for rec in caplog.records
-    )
-
-
 def test_log_position_update_wraps_logging_logger(monkeypatch):
     seen: dict[str, Any] = {}
     original = lp.StructuredLogger
@@ -114,3 +113,67 @@ def test_log_position_update_wraps_logging_logger(monkeypatch):
 
     assert seen["name"] == "position-wrap"
     assert handler.records[0].name == "position-wrap"
+
+
+# ============================================================================
+# Market Data Tests
+# ============================================================================
+
+
+def test_log_market_data_update_supports_optional_fields():
+    with captured_logger("market_data") as (_, handler):
+        lp.log_market_data_update(
+            "BTC",
+            price=Decimal("42000"),
+            volume=Decimal("12.5"),
+            timestamp=123456.0,
+        )
+
+    record = handler.records[0]
+    assert has_attr(record, "operation", "market_data_update")
+    assert has_attr(record, "symbol", "BTC")
+    assert has_attr(record, "price", Decimal("42000"))
+
+
+def test_log_market_data_update_without_optional_fields():
+    with captured_logger("market_data") as (_, handler):
+        lp.log_market_data_update("DOGE", price=Decimal("0.01"))
+
+    record = handler.records[0]
+    assert has_attr(record, "operation", "market_data_update")
+    assert has_attr(record, "symbol", "DOGE")
+    assert not has_attr(record, "volume")
+    assert not has_attr(record, "timestamp")
+
+
+def test_log_market_data_update_with_plain_logger():
+    with captured_logger("market_plain") as (logger, handler):
+        lp.log_market_data_update(
+            "ETH",
+            price=Decimal("1234.56"),
+            volume=Decimal("99.9"),
+            timestamp=987654.0,
+            logger=logger,
+        )
+
+    record = handler.records[0]
+    assert record.name == "market_plain"
+    assert has_attr(record, "symbol", "ETH")
+    assert has_attr(record, "price", Decimal("1234.56"))
+
+
+def test_log_market_data_update_wraps_logging_logger(monkeypatch):
+    seen: dict[str, Any] = {}
+    original = lp.StructuredLogger
+
+    def spy_structured(name: str, component: str | None = None):
+        seen["name"] = name
+        return original(name, component=component)
+
+    monkeypatch.setattr(lp, "StructuredLogger", spy_structured)
+
+    with captured_logger("market-wrap") as (logger, handler):
+        lp.log_market_data_update("SOL", price=Decimal("19.5"), logger=logger)
+
+    assert seen["name"] == "market-wrap"
+    assert handler.records[0].name == "market-wrap"
