--- tests/unit/gpt_trader/utilities/test_logging_patterns_system_health.py
@@ -1,75 +1,0 @@
-from __future__ import annotations
-
-import logging
-from typing import Any
-
-from gpt_trader.utilities import logging_patterns as lp
-from tests.unit.gpt_trader.utilities.logging_patterns_test_helpers import (
-    captured_logger,
-    has_attr,
-)
-
-
-def test_log_system_health_warns_on_non_healthy():
-    with captured_logger("health") as (logger, handler):
-        lp.log_system_health(
-            "degraded",
-            component="monitor",
-            metrics={"error_rate": 0.5},
-            logger=logger,
-        )
-
-    record = handler.records[0]
-    assert record.levelno == logging.WARNING
-    assert has_attr(record, "operation", "health_check")
-    assert has_attr(record, "component", "monitor")
-
-
-def test_log_system_health_wraps_logging_logger(monkeypatch):
-    seen: dict[str, Any] = {}
-    original = lp.StructuredLogger
-
-    class SpyStructured(original):
-        def __init__(self, name: str, component: str | None = None) -> None:
-            super().__init__(name, component=component)
-            seen["name"] = name
-
-        def log(self, level: int, message: str, **context: Any) -> None:  # type: ignore[override]
-            seen["level"] = level
-            seen["context"] = context
-            super().log(level, message, **context)
-
-    monkeypatch.setattr(lp, "StructuredLogger", SpyStructured)
-
-    with captured_logger("health-wrap") as (logger, handler):
-        lp.log_system_health(
-            "degraded",
-            component="execution",
-            metrics={"latency_ms": 120},
-            logger=logger,
-        )
-
-    assert seen["name"] == "health-wrap"
-    assert seen["level"] == logging.WARNING
-    assert seen["context"]["latency_ms"] == 120
-    assert handler.records[0].name == "health-wrap"
-
-
-def test_log_system_health_defaults_to_structured_logger():
-    with captured_logger("health") as (_, handler):
-        lp.log_system_health("healthy")
-
-    record = handler.records[0]
-    assert record.levelno == logging.INFO
-    assert has_attr(record, "status", "healthy")
-
-
-def test_log_system_health_plain_logger_branch(caplog):
-    caplog.set_level(logging.WARNING, logger="health-branch")
-    logger = logging.getLogger("health-branch")
-    lp.log_system_health("degraded", component="risk", metrics={"error_rate": 0.2}, logger=logger)
-
-    assert any(
-        rec.name == "health-branch" and has_attr(rec, "operation", "health_check")
-        for rec in caplog.records
-    )
