+++ var/agents/testing/markers.json
@@ -95,7 +95,7 @@
     ]
   },
   "counts": {
-    "unit": 6152,
+    "unit": 6147,
     "asyncio": 371,
     "property": 82,
     "integration": 79,
